Setting Network ID on UDP (#3675)
* Setting Network ID on UDP

- In order to send/receive network ID to bootnodes, it is necessary to
setup proper network id on udp packet. This PR updates the issue.

* Updated network ID

- updated uint to uint64
- added the comments

+++ cmd/kbn/config.go
@@ -43,7 +43,7 @@ const (
 
 type bootnodeConfig struct {
 	// Parameter variables
-	networkID    uint
+	networkID    uint64
 	addr         string
 	genKeyPath   string
 	nodeKeyFile  string
